+++ crates/client/src/errors.rs
@@ -71,8 +71,6 @@ pub enum AttestationRequestError {
 pub enum ClientError {
     #[error("Substrate: {0}")]
     Substrate(#[from] SubstrateError),
-    #[error("Error relating to private mode")]
-    PrivateMode,
     #[error("Cannot get block number")]
     BlockNumber,
     #[error("Cannot get block hash")]
@@ -147,4 +145,17 @@ pub enum ClientError {
     StripPrefix,
     #[error("subxt rpc error: {0}")]
     SubxtRpcError(#[from] subxt::ext::subxt_rpcs::Error),
+    #[error("Unable to successfully check TDX Quote measurement: {0}")]
+    QuoteMeasurement(#[from] QuoteMeasurementErr),
+}
+
+/// Error when checking quote measurement value
+#[derive(Debug, Error)]
+pub enum QuoteMeasurementErr {
+    #[error("Substrate: {0}")]
+    SubstrateClient(#[from] SubstrateError),
+    #[error("Could not get accepted measurement values from on-chain parameters")]
+    NoMeasurementValues,
+    #[error("Quote verification: {0}")]
+    Kv(#[from] entropy_shared::attestation::VerifyQuoteError),
 }