+++ crates/threshold-signature-server/src/attestation/api.rs
@@ -14,26 +14,17 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 use crate::{
-    attestation::errors::{AttestationErr, QuoteMeasurementErr},
-    chain_api::{
-        entropy::{self, runtime_types::pallet_parameters::SupportedCvmServices},
-        get_api, get_rpc, EntropyConfig,
-    },
-    helpers::substrate::query_chain,
-    AppState, SubxtAccountId32,
+    attestation::errors::AttestationErr,
+    chain_api::{get_api, get_rpc},
+    AppState,
 };
 use axum::{
     extract::{Query, State},
     http::StatusCode,
 };
 use entropy_client::user::request_attestation;
-use entropy_shared::{
-    attestation::{compute_quote_measurement, QuoteContext, QuoteInputData, VerifyQuoteError},
-    BoundedVecEncodedVerifyingKey,
-};
+use entropy_shared::attestation::{QuoteContext, QuoteInputData};
 use serde::Deserialize;
-use subxt::{backend::legacy::LegacyRpcMethods, OnlineClient};
-use tdx_quote::Quote;
 use x25519_dalek::StaticSecret;
 
 /// Retrieve a quote by requesting a nonce from the chain and return the quote in the HTTP response
@@ -64,24 +55,22 @@ pub async fn get_attest(
 #[cfg(not(feature = "production"))]
 pub async fn create_quote(
     nonce: [u8; 32],
-    tss_account: SubxtAccountId32,
+    tss_account: subxt::utils::AccountId32,
     x25519_secret: &StaticSecret,
     context: QuoteContext,
 ) -> Result<Vec<u8>, AttestationErr> {
     use rand::{rngs::StdRng, SeedableRng};
-    use rand_core::OsRng;
-
-    // In the real thing this is the key used in the quoting enclave
-    let signing_key = tdx_quote::SigningKey::random(&mut OsRng);
 
     let public_key = x25519_dalek::PublicKey::from(x25519_secret);
 
-    let input_data =
-        QuoteInputData::new(tss_account.clone(), *public_key.as_bytes(), nonce, context);
+    let input_data = QuoteInputData::new(tss_account.0, *public_key.as_bytes(), nonce, context);
 
     // This is generated deterministically from TSS account id
     let mut pck_seeder = StdRng::from_seed(tss_account.0);
     let pck = tdx_quote::SigningKey::random(&mut pck_seeder);
+
+    // In the real thing this is the key used in the quoting enclave
+    let signing_key = tdx_quote::SigningKey::random(&mut pck_seeder);
 
     let pck_encoded = tdx_quote::encode_verifying_key(pck.verifying_key())?.to_vec();
     let quote = tdx_quote::Quote::mock(signing_key.clone(), pck, input_data.0, pck_encoded)
@@ -116,44 +105,6 @@ pub fn get_measurement_value() -> Result<[u8; 32], AttestationErr> {
     Ok(compute_quote_measurement(&quote))
 }
 
-/// Get our Provisioning Certification Key (PCK)
-/// This generates a quote and gets the public key from it
-#[cfg(feature = "production")]
-pub fn get_pck(
-    _tss_account: SubxtAccountId32,
-) -> Result<BoundedVecEncodedVerifyingKey, AttestationErr> {
-    let quote_raw = configfs_tsm::create_quote([0; 64])
-        .map_err(|e| AttestationErr::QuoteGeneration(format!("{:?}", e)))?;
-    let quote = Quote::from_bytes(&quote_raw)?;
-    let pck = quote.verify().map_err(|e| {
-        AttestationErr::QuoteGeneration(format!("Could not get PCK from quote {:?}", e))
-    })?;
-
-    let pck =
-        BoundedVecEncodedVerifyingKey::try_from(tdx_quote::encode_verifying_key(&pck)?.to_vec())
-            .map_err(|_| AttestationErr::BadVerifyingKeyLength)?;
-    Ok(pck)
-}
-
-/// Get our Provisioning Certification Key (PCK)
-/// In mock mode, this is derived from the TSS account ID
-#[cfg(not(feature = "production"))]
-pub fn get_pck(
-    tss_account: SubxtAccountId32,
-) -> Result<BoundedVecEncodedVerifyingKey, AttestationErr> {
-    use rand::{rngs::StdRng, SeedableRng};
-
-    // This is generated deterministically from TSS account id
-    let mut pck_seeder = StdRng::from_seed(tss_account.0);
-    let pck = tdx_quote::SigningKey::random(&mut pck_seeder);
-
-    let pck = BoundedVecEncodedVerifyingKey::try_from(
-        tdx_quote::encode_verifying_key(pck.verifying_key())?.to_vec(),
-    )
-    .map_err(|_| AttestationErr::BadVerifyingKeyLength)?;
-    Ok(pck)
-}
-
 /// Querystring for the GET `/attest` endpoint
 #[derive(Deserialize)]
 pub struct QuoteContextQuery {
@@ -173,28 +124,4 @@ impl QuoteContextQuery {
             _ => Err(AttestationErr::UnknownContext),
         }
     }
-}
-
-/// Check build-time measurement matches a current-supported release of entropy-tss
-/// This differs slightly from the attestation pallet implementation because here we don't have direct
-/// access to the parameters pallet - we need to make a query
-pub async fn check_quote_measurement(
-    api: &OnlineClient<EntropyConfig>,
-    rpc: &LegacyRpcMethods<EntropyConfig>,
-    quote: &Quote,
-) -> Result<(), QuoteMeasurementErr> {
-    let measurement_value = compute_quote_measurement(quote).to_vec();
-    let query = entropy::storage()
-        .parameters()
-        .accepted_measurement_values(SupportedCvmServices::EntropyTss);
-    let accepted_measurement_values: Vec<_> = query_chain(api, rpc, query, None)
-        .await?
-        .ok_or(QuoteMeasurementErr::NoMeasurementValues)?
-        .into_iter()
-        .map(|v| v.0)
-        .collect();
-    if !accepted_measurement_values.contains(&measurement_value) {
-        return Err(VerifyQuoteError::BadMeasurementValue.into());
-    };
-    Ok(())
 }